producer/requirements.txt
beautifulsoup4==4.12.3
kafka-python-ng
Requests==2.32.3
